+++ Rebus/Bus/Advanced/AdvancedRebusBus.cs
@@ -1,6 +1,11 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using Rebus.Bus.Advanced;
+using Rebus.Extensions;
+using Rebus.Messages;
+using Rebus.Pipeline;
+using Rebus.Transport;
 
 // ReSharper disable CheckNamespace
 
@@ -48,6 +53,40 @@ namespace Rebus.Bus
                 var logicalMessage = CreateMessage(explicitlyRoutedMessage, Operation.Send, optionalHeaders);
 
                 return  _rebusBus.InnerSend(new[] { destinationAddress }, logicalMessage);
+            }
+
+            public async Task Forward(string destinationAddress, Dictionary<string, string> optionalAdditionalHeaders = null)
+            {
+                var transactionContext = AmbientTransactionContext.Current;
+
+                if (transactionContext == null)
+                {
+                    throw new InvalidOperationException("Attempted to forward current transport message, but there was no transaction context!");
+                }
+
+                var currentTransportMessage = transactionContext
+                    .GetOrThrow<IncomingStepContext>(StepContext.StepContextKey)
+                    .Load<TransportMessage>();
+
+                if (currentTransportMessage == null)
+                {
+                    throw new InvalidOperationException("Attempted to forward current transport message, but no transport message was found in the context!");
+                }
+
+                var headers = currentTransportMessage.Headers.Clone();
+
+                if (optionalAdditionalHeaders != null)
+                {
+                    foreach (var kvp in optionalAdditionalHeaders)
+                    {
+                        headers[kvp.Key] = kvp.Value;
+                    }
+                }
+
+                var body = currentTransportMessage.Body;
+                var forwardedTransportMessage = new TransportMessage(headers, body);
+
+                await _rebusBus.SendTransportMessage(destinationAddress, forwardedTransportMessage, transactionContext);
             }
         }