+++ Rebus/Bus/RebusBus.cs
@@ -364,6 +364,11 @@ namespace Rebus.Bus
             await _pipelineInvoker.Invoke(context, _pipeline.SendPipeline());
         }
 
+        async Task SendTransportMessage(string destinationAddress, TransportMessage transportMessage, ITransactionContext transactionContext)
+        {
+            await _transport.Send(destinationAddress, transportMessage, transactionContext);
+        }
+
         bool _disposing;
 
         /// <summary>